docs(asset-meta): rewrite stories in new format (#550)
