--- src/components/AssetMeta/stories/customSpinner.md
@@ -1,22 +1,0 @@
-## Custom Spinner
-
-<!-- STORY -->
-
-#### Usage:
-
-```typescript jsx
-import 'antd/dist/antd.css';
-
-import React from 'react';
-import { AssetMeta } from '@cognite/gearbox';
-
-function ExampleComponent(props) {
-  return (
-   <AssetMeta
-      assetId={123456}
-      customSpinner={<div>Loading...</div>}
-    />
-  );
-
-}
-```
