--- src/components/AssetMeta/stories/selectedPane.md
@@ -1,24 +1,0 @@
-## Returns Selected Pane
-
-<!-- STORY -->
-
-#### Usage:
-
-```typescript jsx
-import 'antd/dist/antd.css';
-
-import React from 'react';
-import { AssetMeta } from '@cognite/gearbox';
-
-function ExampleComponent(props) {
-  const onPaneChange = (pane: string) => { };
-  
-  return (
-    <AssetMeta 
-      assetId={4650652196144007}
-      onPaneChange={onPaneChange} 
-    />
-  );
-  
-}
-```
